fix: update hierarchy

+++ data/tv/src/main/java/com/indisparte/tv/response/EpisodeDTO.kt
@@ -21,7 +21,7 @@ data class EpisodeDTO(
     @SerializedName("still_path")
     val stillPath: String?,
     @SerializedName("vote_average")
-    val voteAverage: Int,
+    val voteAverage: Double,
     @SerializedName("vote_count")
     val voteCount: Int
 )
